refactor(backend): :recycle: new models and relations
now user is the main table and new seeder

+++ backend/models/supply.py
@@ -1,9 +1,9 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, BigInteger, String
 from db.database import base
 
 class Supply(base):
     __tablename__ = 'supply'
-
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
+    
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
     name = Column(String(80), nullable=False)
-    stock = Column(Integer, nullable=False)
+    stock = Column(BigInteger, nullable=False)
